+++ alteia/apis/client/analytics/productsimpl.py
@@ -11,6 +11,7 @@ from alteia.core.resources.analytics.products import (ProductLog,
 from alteia.core.resources.resource import Resource, ResourcesWithTotal
 from alteia.core.resources.utils import search_generator
 from alteia.core.utils.typing import ResourceId
+from alteia.core.utils.utils import parse_timestamp
 
 
 class ProductsImpl:
@@ -163,12 +164,8 @@ class ProductsImpl:
         logs = []
         raw_logs = desc.get('logs')
         for raw_log in sorted(raw_logs, key=lambda log: log.get('timestamp')):
-            timestamp_str = raw_log.get('timestamp')
-            # Ex: '2020-04-16T08:35:42.338000000+00:00'
-            # Truncate before Z chars to keep up to milliseconds
-            z_index = timestamp_str.rfind('Z')
-            clip_index = min(z_index, 26)
-            d = datetime.strptime(timestamp_str[:clip_index], '%Y-%m-%dT%H:%M:%S.%f')
+            timestamp = raw_log.get('timestamp')
+            d = parse_timestamp(timestamp)
             logs.append(ProductLog(timestamp=d, record=raw_log))
 
         return ProductLogsWithTotal(